[Model Monitoring] Move evidently-related code into a dedicated module
To avoid importing it when not necessary.

[ML-8951](https://iguazio.atlassian.net/browse/ML-8951)
