+++ mlrun/model_monitoring/applications/evidently/__init__.py
@@ -13,6 +13,8 @@
 # limitations under the License.
 #
 
-from .base import ModelMonitoringApplicationBase
-from .context import MonitoringApplicationContext
-from .results import ModelMonitoringApplicationMetric, ModelMonitoringApplicationResult
+from .evidently_base import (
+    _HAS_EVIDENTLY,
+    SUPPORTED_EVIDENTLY_VERSION,
+    EvidentlyModelMonitoringApplicationBase,
+)
